refactor: AuthErrorCode 위치 변경
